src/websocket-server/pyproject.toml
[project]
name = "omnibus-ws-server"
version = "0.1.0"
description = "WebSocket server for Omnibus messaging system"
readme = "README.md"
license = "MIT"
requires-python = "~=3.11.0"

authors = [
    {name = "Waterloo Rocketry", email = "contact@waterloorocketry.com"}
]

dependencies = [
    "msgpack>=1.1.0",
    "pyzmq>=27.1.0",
    "python-socketio>=5.11",
    "aiohttp>=3.9.0",
    "omnibus",
]

[tool.uv.sources]
omnibus = { workspace = true }
